add gl3d plot index + register it in plotly.js + remove Gl3dLayout
